+++ app/api/v1/groups/[id]/actions/index.ts
@@ -12,3 +12,5 @@ export {
   handleAddMember,
   handleRemoveMember,
 } from './group-crud';
+
+export { handleGetState, handleSetState, handleResetState } from './state';